app/requirements.txt
setuptools>=69.0.3
wheel>=0.42.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
passlib==1.7.4
bcrypt<4.0
python-multipart==0.0.6
pandas>=2.2.0
numpy>=1.26.0
joblib==1.3.2
ta==0.10.2
tinkoff-investments==0.2.0b113
python-dotenv==1.0.0
requests==2.31.0
apimoex==1.3.0
catboost
pytest
pytest-asyncio
httpx>=0.25